Rework HonkBonk.py
Removes and simplifies much of the (somewhat bloated) foundation of the bot.
Many methods have been moved to helpers.py
Scheduler has been moved to its own module.
Niche or outdated methods and attributes have been removed.

The intent of these changes is to refine what was good and remove what was bad for further development.